Keep (better) track of what we're doing via Todot
Okay, so previously I had my to-do list in a text file, sitting around
in my project's root directory. This was meant to be more of a
long-term to-do instead of an immediate to do list.

That worked pretty well, but presented some issues. First, since
looking at the to-do wasn't part of my regular workflow, the document
was prone to drifting from my actual development goals. Secondly,
since I had no way of tracking my moment-to-moment goals, I often had
to spend an hour or so reorienting myself as to my current goal.

Fortunately, I found the Todot plugin. This plugin adds a small Todo
widget in the Godot editor. This allows me to track tasks locally. So
I moved the old to-do text into this new tracking task. Seems to work
well so far!