Solidfy animation hints and demands
While developing attacks, I realized that we had no way of asking for
a specific animation to be played. So if we needed something like an
attack, we had no way of forcing that animation to play, nor any way
of knowing when that animation was complete.

So, I've added in new signals and functions relating to animation. The
demand_animation and demand_complete signals allow us to ask for
animations and be notfied when they're completed. The DemandAnimation
task acts as an action wrapper for demanding animations.

To reflect the new way we handle animations, I tried to rename the
different Pawn frames using the KMAD scheme: KEYWORD, MOVEMENT,
ATTITUDE and DIRECTION. The result seems to work pretty well, but I
think we've gone as far as we can with the current admixture of test
assets. I think we need to retool the old Pawns to more closely matchup
with the newer Pawns.