Add options to disable TCM wandering and cowardice
One of the things that bugged me about the TCM was that the Fleeing
state could overwrite the current task, and it's habit of moving
around when idle.

Of course, I TOLD it to do those things, but these things I TOLD it
to do were starting to get a little under my skin. It made them
difficult to work with as far as "testbeds" were concerned.

So, I've added toggles to the TCM for enabling the wandering and
fleeing behaviors.