Add a prototype melee attack
Because attacks are - in most games - tied very intimately into
animations, we cooked up the demand animation system. And that worked
pretty well, so we just needed to add the actual attack part... right?

Well, turns out that part isn't quite simple. Originally I had dreamed
that we would have an "attack" action that just did the attack on an
entrance. The task chain would go Demand -> Attack -> Demand. However,
there is a significant (or rather, not insignificant) delay when
switching between states. This is normally fine. However, there will
be range restrictions for attacks, so time & speed are of maximum
importance.

Ergo, I cooked up a more complex action called MeleeAttack. It can
perform a lead-in animation, make the attack, then do a lead-out
animation all in a single action. It handles the logic for there being
no lead-in or no lead-out. It handles everything!

Except actually doing damage. Right now it just reports that an attack
was made. Guess I need a way to determine what I hit, or tell
something that it got hit...