Pass damage to the appropriate object when hit
I closed off the last commit having created an attack but not having
any way to communicate that to whatever was being hit. I did that
because, honestly, I assumed it would be hard.

We had a numerical object ID and I assumed there would be some whole
giant litigated process to getting the exact object in question. Turns
out, it was just a single function call. Isn't that convenient!

So, we now use the standard 'take_damage' method to send some damage
to the object. Works well enough; we'll need to develop more robust
systems as time goes on, of course.