Add a scene that creates 3D Sprite text
Sometimes we need to display text in the game's 3D space, not just
the GUI. We tried to do something about it with the "3D GUI" scenes,
which ended up just being a QuadLabel.

QuadLabel was nice but had some problems. Firstly, the textures had
to be instantiated in a weird way since each label technically
constituted it's own screen. This would also lead to painful setup if
there were multiple instances of a QuadLabel in the scene. There was
also inconsistent scaling issues, and the GUI was frequently not lined
up correctly within the label.

To get around this, I've created a DynamicSpriteText scene. It
requires a spritesheet in a very particular format, but it's much
easier to work with than the QuadLabel scene.

With respect to that funky format: it's 95 characters in an exact
sequence all on one line. I chose that because all the bitmap fonts I
had at the time used this format, though it looks like others don't
quite work the same way...