Replace the QuadLabel with DynamicSpriteText
I was very happy with how the DyncamicSpriteText scene turned out.
Much happier than I was with the broken QuadLabel scene. So I made
moves to replace it.

It pretty much was a smooth swap out. I updated some of the
documentation and I'm pretty happy with how everything's going.