Add floating, self-destroying text messages
A pretty common "effect", or should I say "visual", is what I'll call
"float text". It's text that is usually associated with a point on
screen and is meant to inform about something that happened at that
point on the screen. The best example I can think of would be damage
indicators in certain games, most typically fighting games. I've also
remember it being somewhere in Zoo Tycoon...

Anyway, it's a perfect way to communicate information at a certain
point in time about a certain point of space. So, I decide to whip up
my own version of this concept using the DynamicSpriteText scene.

It's scaled a bit up, and has some variables to control lifespan and
speed. Other than that, it's more or less the same scene. I stuck it
in the Pawn so we can see damage values. It works pretty well! I think
that I'll eventually use this for other things. This makes a pretty
good argument for having a "MasterMotionAI" scene that can handle this
similarly to any and all other damage in a common way.

Now, to make this damage actually mean something...