Add rudimentary damage taking and death
Previously we were just printing out values for damage - it wasn't
actually DOING anything.

So I added in code to pass the damage to the CharacterStatsCore, which
handles the damage and informs us if the character is supposed to be
dead or not. I left in some blank variables there to deal with
different types of damage.

Unfortunately this caused a crash because of how we're spawning damage
floats, so I had to comment that out. It occurs to me now that we're
gonna need to overhaul the health and damage so that all kinds of
different nodes (i.e. structures, items, and motion ai) handle damage
similarly. This is gonna be a tricky one for sure...