Implement a common stats core for physics objects
As we were working taking damage and dying, I realized that more than
just the Motion AI was going to need to take damage. We'll need items
to take damage, and we'll need structures to take damage.

But our core was designed exclusively for Motion AI. We needed a new
core that could be common to Motion AI, items, and structures.

Enter the CommonStatsCore! It's a micro-ized version of the
CharacterStatsCore, shaved down to just the health and damage
components. It also has code for showing damage values, so we won't
need to write unique code for each item, structure, or Motion AI.

The CharacterStatsCore has been changed to derive from the
CommonStatsCore. The "generalization" process meant we had to change
some things - for example "character_died" became "object_died". Even
though it doesn't make sense that a door can 'die', we going to
consider doors and items as things that can die.

I also updated the documentation appropriately.