Add respawning targets to the Attack Model Test
Previously, we could attack the target pawn of the Attack Model Test
once or twice and then they would die. Then if we went to make another
attack, the program crashed. And that was it!

Of course that wasn't a very useful test. So, I've changed how the
test is run. Now, when the Pawn dies, it emits a signal. This signal
is intercepted by the test, which spawns a new Pawn and restarts the
looping cycle.

And it works pretty good. Yipee!!!