Update XSM to 1.4.6 and set new targets
I spent some time thinking about our next steps, and thusly updating
our to-do list.

We need to work on how the AI targets moving objects. Right now it
constantly chases its target, instead of intercepting it, because it's
always moving to where the target WAS, not where the target is GOING.
That's going to require some reworking, I think...

Also, I've spotted a potential bug: the current code measures our
distance to our target to identify when the moving object is stuck.
However, if our target is moving, this value will change regardless of
whether we are actually stuck or not. So we'll probably need to rework
how the movement code detects a "stuck". It should probably be some
measure of distance to the previous position. Or something.

And while we're at it, we should probably generalize the BasePawn into
a base MotionAI class. Or something.

Anyway, I also took this opportunity to update XSM to version 1.4.6,
which seems to be pretty stable.