Add movement-predicition to the Falling Body Machine
We're currently trying to add some sort of movement prediction vector
to the different machines. I figured it'd be easiest to start with the
absolute simplest machine - the Falling Body Machine.

The overall architecture is as such: when we enter the state machine,
we clear out the prediction vector. It is then up to the child states
to modify the prediction vector accordingly. It works well for the
Falling Body Machine, but of course that's also the absolute simplest
machine possible.

I also added in an "OnGround" state unique to the Falling Body, since
we were just using the Kinematic Driver Machine's OnGround state.