Add movement predictives the Kinematic Machines
The next step on our movement-predictive development path was adding
the predictive into the Kinematic Driver Machine.

It took some puzzling to figure out where and when to best update the
movement predictives - I ultimately just went with 'as late as
possible'. The movement processing is already more than complicated
enough, it made the most sense to just tack it on at the end, changing
the values depending on the context.

I also added in some debug drawery, so that I could visualize the
projected movement. This helped me realize something: this movement
projection vector is going to be UTTERLY USELESS! IT CHANGES
CONSTANTLY! IT'S HIGHLY INACCURATE!

A short search revealed that, apparently, "intediction pathing" (which
I was trying to do) is a WHOLE OTHER REALM of algorithms and MATH. So
we might table our dreams of interdiction for a hot second...