Update the REM and TCM with the movement projective
The overall structure of the Rat Emulation Machine & Tasking Coward
Machine's Physics Travel Region (PTR) is almost an exact duplicate of
the Kinematic Driver Machine. Ergo, it was a natural next step to
migrate some of those changes over.

There were some small differences - for example, what the 'target'
variable refers to - but other than that it was a rather seamless
migration.

I did run into some errors with the Walk states - all three of them.
Apparently it (was) POSSIBLE to reach the movement-projection stage
with no target position to move to. This caused a fault when we tried
to subtract a Vector3 from a null. Changing to an "elif" from an
"else" fixed the issue.