Add a base status condition
I'm a bit scatterbrained when it comes to this project and I'm not
quite sure what I was doing. And for some reason or another, I decided
I wanted to work on status conditions.

Before defining basic status conditions, I wanted to work on their
commonalities. So I created a basic status condition that all future
status conditions will derive from.

Each will have a number of modifiers - composite array-objects with
different fields. This will standardize how modifiers are described
so we can (hopefully?) do some automatically generated explanations
later on.

The fields modified are stored as strings that we'll check against.
This loose coupling between the status conditions and the status core
fields will hopefully allow us to be more dynamic in what fields are
and aren't affected (hopefully).