Add some a basic fire particle effect
One of the reasons I stalled out on status effects was because I
couldn't think of a way to communicate the status effects. Like, fire.
How do we do fire? How does that even work?

After some messing around with third party tools, I decided that, for
now - and we really may need to change our mind on this in the
future - we're just going to rely on particle effects. Those seem to
look the best, and the current performance impact is... negligible?
We'll have to see.

So, I've added a new particles subdirectory to our special effects
directory. I added a Particles scene called 'Fire32' (because the
particles are 32x32 diamonds). I also added a 'ParticleZooTest',
which is made to test performance impacts and visual movement effects.

Now, I wonder if I can whip up another particle effect real quick...