Add a global "Manifest" scene
We're going to add a whole bunch of different particle effects over
time, so I was going to expand the Particle Zoo test to cover multiple
different particle materials.

However, before I sat down to add a filelist functionality to the
Particle test, I remembered something that's been gestating at the
back of my mind - there's so many instances where we need to list out
spawnable assets. You have the tests, there's the particle effects,
and eventually we're gonna add lists for different spawnable motion
AI.

Now a key issue is one of mods - if we load a mod that adds a new type
of NPC, where do we define the existence of that NPC. Where do we list
it?

That's what the new Manifest autoload is for. It's our one stop shop
for lists of lists. It can be accessed and modified globally. I'm not
gonna lie, that sets off different sorts of alarm bells for me. But
I'm SUUUUUURE it'll be fiiiiiiiiiine.