Describe emitters with Rich Particle Materials
We want to dynamically create, and remove, particle emitters from NPCs
as they gain and lose status effects. However, the configuration of
particles is split up into the emitter and the material. This can be
problematic, as the material defines what happens to the particles
(color changes, scale changes, emitting area, etc) whereas the emitter
controls how many particles are spawned and WHAT THEY LOOK LIKE. Not
great for us.

To get around that, I've created the RichParticleMaterial, which
basically defines everything we need to know about an emitter and
packs it into the Material resource. From this, we'll be able to
create emitters on the fly. Now I just have to do that...