Create emitters quickly with Rich Particle Emitter
As outlined previously, we're trying to decouple the particle
materials from the emitters so we can easily scale different particle
effects up and down - depending on the size of whatever is effected.

So, I've created a Rich Particle Emitter to couple with the Rich
Particle Materials. The result is... glitchy. I'm not gonna lie,
normally I wouldn't commit code like this but I need to capture this
moment in time so we can return to it.

Anyway, the actual particle set-up seems to be working fine; what's
actually breaking is the AABB calculation - which just so happens to
be 90% of the code. That's okay, I think I can work with that...