Ensure particles maintain their size regardless of scale
Previously, scaling the particle systems up and down would also scale
the particles up and down. In most situations that would be ideal BUT
the way we want to use the Rich Particle Emitter is to scale the
emitters up and down to control the emission area.

So, we had to use MATH to create NEW MESH SHAPES that were THE
APPROPRIATE SIZE FOR A GIVEN SCALE. I'll admit, I'm a bit worried
about the possible performance impacts of spawning new meshes
everytime we initialize a particle effect, but I thiiiiiiiink Godot
will be able to tank it. We just need to be careful about
when-and-where we spawn them.

The AABB's are STARTING to look better, but they're not quite there
yet. Think I need to do some more work...