Dynamically scale AABB to match particle movements (mostly)
After some tinkering, I've gotten the AABB scaling to be, like, 90%
correct. There were some typos with how I was setting the maximum &
minimum, especially & explicitly for Y.

I was also scaling things wrongly. Since the scaling takes care of
most things, I just needed to leave it out of the initial spawn-box
length calculations.

And it seems to work? There are some rogue particles escaping to the
sides but everything seems to be in working order.