Test any MANIFEST'd particle with the Particle Zoo
Previously, we were hard coding the different particle materials into
the particle zoo. If we wanted to change something, I had to manually
adjust the gd-script.

With this update, the Particle Zoo Test is now much more dynamic: you
can manually specify which Particle Materials to use (from a MANIFEST
backed, dynamically filled list), set the scale to use for the
particle systems, and start-stop the tests.

This has been really useful, and I've already made some good
observations. For example: the density system is completely unhinged.
I need to fix that.