Scale particle count by cubic-rooting the volume
Okay so I've been tussling with how to scale the particle counts for
my different particle systems. Originally, we had a DENSITY value, and
that density value was multiplied by the volume to scale the number of
particles coming from a node appropriately.

But it was growing too big, too quickly. I had no idea how to control
that growth!

So, I sat down and thought: the reason it's growing out of control is
because volume is cubic. It's inherently an exponential function. You
always multiply three sides and that's why it grows so quickly. If we
wanted some sort of sane scaling, I'd either need to find a way to
make the cubic function strictly linear or find a different value.

The I realized: If my problem is that I'm calculating x^3, and I just
wanted x, then the solution was to take the cubic root!

So, the "particle density" is now the "Root Particle Slope". As
stated, it's the slope that gets multiplied by cubic root of the
volume. It actually works REALLY well and results in particle systems
that look consistent regardless of the scale.