Cook up some "slash" effects
Played around with some different "slash" effects; these will be the
particle effects palced on a body when it gets hit by an attack.

They look okay; honestly they don't look so great in the zoo but
that's also not where it's meant to shine. I think it'll work for now.

I did have some trouble with editing the Rich Particle Materials;
since they're all supposed to behave the same but have different
shapes, that meant I had to manually move different settings across
the various materials to coordinate. Definitely one of the key flaws
of my Rich Particle idea...