Update XSM to 1.5.4
I've had an extended hiatus - I was doing a lot. Painting 40k Kill
Team models, writing Monster-of-the-Week episodes, absolutely LOSING
myself to Elden Ring. But it's always good to come back to
programming, which is always my favorite.

Anyway, such a long hiatus means I've kinda lost track of what I was
doing. But there's always maintenance of a sort to be done! First up,
our plugins.

The previous XSM version was lingering around 1.4.6, but had since
been updated to 1.5.4 (1.5.5 on the GitLab version). The update
doesn't seem to have caused any issues, so now we need to take a look
at the other plug-ins...