Update Qodot to 1.7.1
Qodot, which is really a bedrock plugin for this project, needed to be
updated to version 1.7.1. Updating Qodot has been pretty troublesome
in the past, but fortunately it was smooth this time. I think it was
painful previously because I may have been a bit... "overzealous" in
deleting and trimming files. I was much more reserved this time, so
the update was far easier.