Separate particle visuals from behavior as best we can
One of my biggest issues with the RichParticleEmitter was how closely
we had coupled together the visuals of the particles - the sprite, the
draw passes - with the behavior of the particles - i.e. everything and
everything in the particle materials.

Now this was done on-purpose, accidentally. I'm not really sure how to
term it. See, to scale the particle systems up-and-down, we needed a
consistent set of information. I, thinking a little uncritically, just
rolled this all up into the RichParticleMaterial.

However, this started to create problems for us. Things were fine
initially, but we ran into trouble when we created all of our hit
particles. We had a consistent behavior that I really would've liked
duplicated for all the hit particles, but I couldn't! It all had to be
tied up in the RichParticleMaterial.

So, I've redone everything - it's similar, yet different. We now have
separate spatial and particle materials. The particle materials are
normal, while the spatial materials have been augmented into
ParticleReadySpatialMaterial (bit of a wordy name but I couldn't
immediately think of anything better). These two are then bound
together in the ScalableParticleBlueprint, which includes particle
system-specific configurables which I couldn't quite place on either
the particle or spatial materials. This blueprint then takes the place
of the RichParticleMaterial. The RichParticleEmitter has thusly been
reworked into the ScalableParticleEmitter, with minimal notable
changes.

Now then, as much as I love my scalable particles, I also want to
support prebuilt particle systems. I also need to redo the status
effect / particle interface, and then I need to sit down and redo
attacks with some sorta... fancy... scanning... something.