Allow ShaderMaterials with Scalable Particle Emitters
Okay, I spent way too much time making a stupid helix particle effect.
And then, I thought to myself, "Oh, that wouldn't actually be that
hard to scale up using the ScalingParticleEmitter, if we assume the
1x1x1 volume restriction".

Anyway, I then spent too much time making that work, and as much as I
am angry at myself and the ShaderMaterial for taking up my time with a
stupid diversion, it seems to work pretty seamlessly.

Still, this is more cavalier than I usually go about things, and
overall I feel like it's tacked-on or jury-rigged. I guess that's
because all I really did was make an algorithm harder to read by
adding a bunch of if statements? Yeah, that's probably it.

I updated the documentation to match my newest improvements - I
noticed a bunch of errors and half-finished thoughts, so I corrected
those. Of course, I thought the documentation was fine last time, so
who knows. Zarro boogs!