Use an inner class for status effect modifiers
Previously, we were defining modifiers using arrays, with each index
of the array having a semantic meaning. We did that because, as far as
I was aware, classes in Godot were either resources or nodes. There
was no such thing as a stand-alone inner class.

Until, that is, I accidentally typed out "class" instead of
"class_name" and found out that apparently you CAN declare an inner
class in Godot. I feel like that was an update that I missed at some
point.

Anyway, the class was CLEARLY the better way to go about it, so I've
changed everything around to make use of the new StatMod inner class.
And I'm pretty happy with it! I also updated the documentation, since
that was really out-of-date.

Now, I think I could do better: what if, instead of the StatOp enum,
we had multiple kinds of StatMod that had the appropriate fields, and
also had a function that could modify a given core on demand? Or
something like that...