Scale up ShaderMaterials in the Scalable Particle Emitter
I was testing out the particle systems, but I couldn't get my singular
shader scalable blueprint to actually scale up. The amount of
particles was always fixed.

Turns out, I never actually implemented the code that set the amount
of particles for the ScalableParticle emitter. I set the spawn lengths
for each axis and then never bothered calculating the volume or
scaling off of that. Whoops.

I used the volume formula for a standard box, since we assume that the
base particle system is contained in a 1x1x1 box.