Add a Condition Zoo
I was trying to test out status conditions when I realized that I
didn't really have a good venue for doing that. No way to observe
their effects.

So, I added a Condition Zoo - I derived it from the AttackModelTest
and based the design on the ParticleZooTest. It's pretty good, I
think.

Of course, this meant that I had to actually implement the status
effect add-and-remove from the CommonStatsCore.