Implement status effect damage-over-time and lifetime
I was working on a fire status effect when I realized that I still
hadn't implemented damage or status effect lifetimes. I had fields for
them but hadn't actually bothered implementing anything.

The first thing I changed as part of the implementation: I was going
to spawn in Timer nodes for the lifetime and dot_interval variables,
when I realized that spawning in Timer nodes was very un-Godot.
Instead, it'd be better if the timers were already there and acted as
our lifetime & interval configurables. So, the lifetime and
dot_interval have been removed - effectively replaced by two timers
and two corresponding enable/disable booleans.

That timer rework/change constituted the lion's share of the changes.
I also added signals to the status effect, which are now registered in
the base status condition node. I also added that "fire" status effect
and tested it.

Now then, I think I can do something better with how we're spawning in
the particle effects...