Spawn particles under the status conditions
Previously, we spawned particles in the CommonStatsCore and then
manually put them under the appropriate status condition. After a time
I realized that was kind of pointless - why couldn't the condition
just spawn it's own particle effects?

There were some reasons for why we did it the old way. Firstly, the
BaseStatusCondition was meant to just be a data class - the sort that
I'd call a "Model". It wasn't meant to have functions, really. It
wasn't meant to do anything outside of itself. But other things forced
us away from that architecture, so there was no sense in being a
stickler for it.

The other reason was that the CommonStatsCore will have our system
scale information, so it made sense to spawn it where that information
was. But not too much sense. This new way is cleaner, and we can
control the "scale" using an optional argument.

All that considered, we now use the spawn_particles() function to
create the particles as appropriate. Now, I need to decide what to do
next. I'm thinking either a healing status effect, or creating a
"BaseAgent" class of some kind, to standardize how we hit items and
NPCs in this game...